refacor: 라벨 서비스 로직 리팩토링
try-catch 문 삭제 ( controller에서 에러 확인)
필수적 입력갑 체크
update, delete시 0이 반환되면 error 메시지 반환

+++ backend/src/service/label.js
@@ -2,33 +2,35 @@ const Label = require('../model').Label;
 
 module.exports = {
   create: async ({ color, title, content }) => {
-    try {
-      return await Label.create({ color, title, content });
-    } catch (error) {
-      return { error };
+    if (!color || !title) {
+      return { error: '정보가 부족합니다' };
     }
+    return await Label.create({ color, title, content });
   },
   read: async () => {
-    try {
-      return await Label.findAll();
-    } catch (error) {
-      return { error };
-    }
+    return await Label.findAll();
   },
-  update: async ({ id, color, title, content }) => {
-    try {
-      await Label.update({ color, title, content }, { where: { id } });
-      return true;
-    } catch (error) {
-      return { error };
+  update: async ({ id, data: { color, title, content } }) => {
+    if (!id || !color || !title) {
+      return { error: '정보가 부족합니다' };
     }
+    const [label] = await Label.update(
+      { color, title, content },
+      { where: { id } }
+    );
+    if (label) {
+      return true;
+    }
+    return { error: '없는 id값 입니다' };
   },
   remove: async ({ id }) => {
-    try {
-      await Label.destroy({ where: { id } });
-      return true;
-    } catch (error) {
-      return { error };
+    if (!id) {
+      return { error: '없는 id값 입니다' };
     }
+    const label = await Label.destroy({ where: { id } });
+    if (label) {
+      return true;
+    }
+    return { error: '없는 id값 입니다' };
   },
 };
